refactor: inline extension classes into related source files
