+++ src/SharpProp/IJsonable.cs
@@ -12,3 +12,17 @@ public interface IJsonable
     /// <returns>A JSON string.</returns>
     string AsJson(bool indented = true);
 }
+
+internal static class JsonExtensions
+{
+    private static readonly JsonSerializerSettings Settings = new()
+    {
+        Converters = [new StringEnumConverter(), new UnitsNetIQuantityJsonConverter()],
+    };
+
+    public static string ConvertToJson(this object? instance, bool indented)
+    {
+        Settings.Formatting = indented ? Formatting.Indented : Formatting.None;
+        return JsonConvert.SerializeObject(instance, Settings);
+    }
+}
